The Exception classes was removed to the CBasicReport.h file
