+++ Common/Core/CBasicReport.cpp
@@ -19,7 +19,6 @@
 #include <QMetaObject>
 #include <QMetaClassInfo>
 #include "elcUtils.h"
-#include "MMSTypes.h"
 
 const QChar eolChar(QLatin1Char('\n'));
 const QChar emptyChar = QChar();
@@ -101,7 +100,7 @@ CBasicReport::setReportDataItem(QXlsx::Document *report, const int dbFieldIndex,
 {
     QVariant writeValue = m_db->geValue(dbFieldIndex);
     if (!report->write(row, reportFieldIndex, writeValue))  {
-        throw mms::XlsxError();
+        throw XlsxError();
     }
 }
 
@@ -110,7 +109,7 @@ CBasicReport::setReportDataItem(QXlsx::Document *report, const QString &dbFieldN
 {
     QVariant writeValue = m_db->geValue(dbFieldName);
     if (!report->write(row, reportFieldIndex, writeValue)) {
-        throw mms::XlsxError();
+        throw XlsxError();
     }
 }
 
@@ -118,7 +117,7 @@ void
 CBasicReport::setReportDataItem(QXlsx::Document *report, const int column, const int row, const QVariant &writeValue)
 {
     if (!report->write(row, column, writeValue)) {
-        throw mms::XlsxError();
+        throw XlsxError();
     }
 }
