+++ Common/Core/CEventLogReport.cpp
@@ -19,7 +19,6 @@
 #include <QScopedPointer>
 
 #include "DBStrings.h"
-#include "MMSTypes.h"
 
 CEventLogReport::CEventLogReport(QObject *parent)
     : CBasicReport{parent}
@@ -81,7 +80,7 @@ CEventLogReport::generateReport()
 
             writeValue = m_db->geValue("timestamp").toDateTime();
             if (!xlsxReport->write(row, colTimestamp, writeValue, dateFormat)) {
-                throw mms::XlsxError();
+                throw XlsxError();
             }
             setReportDataItem(xlsxReport.data(), "externalip", colExternalIP, row);
             setReportDataItem(xlsxReport.data(), "username", colUsername, row);
@@ -89,7 +88,7 @@ CEventLogReport::generateReport()
 
             writeValue = checkDetails(m_db->geValue("details").toString(), isDataTooLong);
             if (!xlsxReport->write(row, colDetails, writeValue) && !isDataTooLong) {
-                throw mms::XlsxError();
+                throw XlsxError();
             }
             setReportDataItem(xlsxReport.data(), "authtype", colAuthType, row);
             setReportDataItem(xlsxReport.data(), "internalip", colInternalIP, row);
@@ -100,7 +99,7 @@ CEventLogReport::generateReport()
                 break;
             }
         } // while
-    } catch (mms::XlsxError &ex) {
+    } catch (XlsxError &ex) {
         setErrorString(ex.what());
         retVal = false;
     }
